+++ Expanded Cheat Menu/Code/Menus/DebugFunc.lua
@@ -483,7 +483,7 @@ function ChoGGi_Funcs.Menus.ExamineObject()
 		-- off centre of central monitor
 		local width = (terminal.desktop.measure_width or 1920) - (dlg.dialog_width_scaled + 100)
 		dlg:SetPos(point(width, 100))
-		ChoGGi_Funcs.Common.ToggleConsole(true)
+		ChoGGi_Funcs.Common.ShowConsole(true)
 	end
 end