+++ Mods ChoGGi/Example New Nation/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "New Nation Example",
+	"id", "ChoGGi_NewNationExample",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[
+Adds a fake nation and some names to be used with a sponsor.
+
+]],
+})