+++ Mods ChoGGi/Orbital Prefab Drops/items.lua
@@ -1,6 +1,13 @@
 -- See LICENSE for terms
 
 return {
+	PlaceObj("ModItemOptionToggle", {
+		"name", "EnableMod",
+		"DisplayName", T(302535920011303, "<color ChoGGi_yellow>Enable Mod</color>"),
+		"Help", T(302535920011793, "Disable mod without having to see missing mod msg."),
+		"DefaultValue", true,
+	}),
+
 	PlaceObj("ModItemOptionNumber", {
 		"name", "ModelType",
 		"DisplayName", T(302535920011457, "Model Type"),